vbdev_compress: allow setting chunk_size during bdev_compress_create
Chunk size is the unit size for every compression operation.  It was
currently hardcoded to a default of 16k so this adds a -c parameter to
bdev_compress_create to specify a different size. To make sure that
valid values are tested, this is currently restricted to 16k or 32k.

If this was to be expanded in the future, note that the following must
be respected:
* chunk_size must be an even multiple of backing_io_unit_size and
  logical_block_size (as described on include/spdk/reduce.h)
* chunk_size must be at most (64k - 1) because an rte_mbuf max size is
  16 bits. Going above 64k is possible but would require some
  refactoring to chain rte_mbufs so that neither is larger than 16 bits.

Also add the chunk_size to the output of bdev_get_bdevs (and while we're
at it, add lb_size and backing_io_unit_size as well).

Finally, update the documentation and tests. And while we're at it, set
lb_size to required.

Signed-off-by: Martin Oliveira <martin.oliveira@eideticom.com>
Change-Id: I9b8a84760e1204235c2f2554edb259d7a3315ed5