lib/idxd: Increase max channels per device
SPDK uses only one work queue to submit work
to a DSA device. Since the DSA portal address
space and number of work queue slots are limited,
there is a limitation for the number of concurrent
I/O channels which can be used with a single
DSA device.

This patch increases this limit from 8 to 32,
since the previous value might have been,
in some cases, too low.

Change-Id: Ia1d6cf936dc1a3c826009877458a80b8a7edba8d
Signed-off-by: Slawomir Ptak <slawomir.ptak@intel.com>