go/rpc: Initial implementation of rpc call generator
This is the first iteration of struct generator for JSON RPC calls.
Currently, in SPDK, we have three places (four with Go) that
have to be updated with addition of new RPC call. To remedy this issue
we want to have one (two at most) places to update with new RPC calls.
This patch contains code for structure generation for Go. It uses JSON
schema file that contains RPC calls.

Change-Id: Ib7b8c6cd334ccc11ca18b91670d482f5b7cc71fc
Signed-off-by: Maciej Mis <maciej.mis@intel.com>