module/keyring: add linux keyring module
This module uses the keyutils library to fetch keys managed by the Linux
kernel keyring.  The keys are created implicitly in the probe_key()
callback.  For now, they need to be attached to the kernel's keyring
before requesting them.  The following patches will add the ability to
specify callout information which will allow  the kernel to instantiate
the keys if they're unavailable.

Change-Id: I0aedd19247c6c35f22bcb61f4b387d5d85b42cdf
Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>