lib/idxd: assign specific portal offset to each descriptor
Each channel gets predefined number of descriptors and portal addresses,
it can write to. Portal offset changed on each submission.
Number of descriptors for each channel was limited by chan_per_device.

Rather than calculating on the I/O path, assign unique portal
address to each descriptor based on an offset from portal address.
This will allow for arbitrary number of descriptors for each channel
in the future.

Please note that some of the descriptors will be assigned
duplicate offsets to the portal. For example due to
WQ size being 128 descriptors, portal size 4096 bytes,
and each descriptor 64 bytes - resulting in 2 descriptors
pointing to the same offset.

Next patch adds another field to track offset and clear bit
in the array separately in the ops. As to not go over the
the cache line in idxd_ops, uint16_t is used for both.
With WQ size of 128 and portal size 4096, this should still fit.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Change-Id: Idcf5ccdf5b76fa58c15a6c3d160d857e65bf4303