keyring/linux: add callout info
Users can now specify callout info that will be passed to the kernel in
request_key(2).  If probing for a key that isn't already in the keyring,
the kernel will try to instantiate it by executing the /sbin/request-key
program which in turn runs user-defined application (configured as
described in the request-key.conf(5) man page) passing the callout info
as one of its arguments.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I0b1737d0449087f3ba0e23fc23073d5612e7cb8c