nvme/pcie: mark qpair as ready after its fully initialized
The completion for the CREATE_IO_SQ command can executed on any thread,
as it's an admin command.  If it's executed on a different thread than
the one polling the IO qpair, setting shadow doorbell addresses and
nvme_pcie_qpair_reset() is racing with with
spdk_nvme_qpair_process_completions() called on that thread.  Of course,
it's probably extremely unlikely to encounter this in reality, as it'd
require two process_completions() calls to happen after setting qpair's
pcie_state to QPAIR_READY, but before executing the rest of the
initialization, but it's still a data race, so it needs to fixed.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I718485216a17ab84ccf1838add62f3eadc5170fa