test/nvme/sw_hotplug.sh: return unused drives to nvme driver
sw_hotplug.sh was limited by this commit: 7278ac9 to run on at most 2
nvmes. Since hotplug's spdk_nvme_probe() call is not provided with
a trid, it checks for hotplug events on all devices in userspace.
We should return unused nvmes to the original driver to avoid
detecting events not triggered by sw_hotplug.sh.

Change-Id: Iabe38a57c7cdfec6ca28eb27fb41eac71402fb88
Signed-off-by: Krzysztof Karas <krzysztof.karas@intel.com>