bdev: [RFC] Add a per-IO action flag
The idea is to include some level of IO buffers pre-processing in the bdev
library. It would include copying the data, leaving space for the metadata
and inserting/stripping DIFs.

The original idea was to implement all that functionalities in a new
vbdev DIF layer. This is useful for the applications, but there are
cases where SPDK generates some internal IO. For bdevs with metadata
(extended sector size or separate) without DIF, or with DIF, all of this
code needs to account for these cases: leave space for the metadata,
allocate a separate metadata buffer, copy the data around, etc. If we
go the vbdev DIF route, there is no easy way to use it there.

The internal IO can be generated by:
- RAID superblock
- blobstore metadata
- vbdev GPT
- FTL
- Any libs/modules added in the future

The proposed idea is to include that logic as an extension to
the existing bdev read/write functions. This way the vbdev DIF
is no longer needed, and all the internal IO can easily use it.

The bdev lib would maintain a pool of additional buffers used
when data copy is required, and would use the Accel API to
generate/strip DIF/DIX.

Change-Id: Iddfec48a11658e7db7c738363723e65043c29fb1
Signed-off-by: Slawomir Ptak <slawomir.ptak@intel.com>