lib/nvmf: suppress memory leak detection for ctrlr struct
LSAN sometimes detects false memory leaks when allocating
memory for ctrlr structure, probably due to delayed free operation.
It was verified couple of times there is no memory leak here,
and the memory is freed as expected.

This patch suppresses memory leak detection for this
allocation to enable long fuzzer job in Jenkins.

Change-Id: I86cc116d9f49fe27740f5bd8f109c97dc6b330c7
Signed-off-by: Marcin Spiewak <marcin.spiewak@intel.com>