module/sock: do not send new session tickets
According to RFC 8446 section 4.6.1:
   At any time after the server has received the client Finished
   message, it MAY send a NewSessionTicket message.  This message
   creates a unique association between the ticket value and a secret
   PSK derived from the resumption master secret."
This means that it is not required to send NewSessionTicket messages.

The original behavior (using OpenSSL default number of tickets) is
preserved and is the default setting.

Kernel side is not yet prepared to process these tickets after
successful handshake and will abort TLS connection upon encountering
that type of post-handshake message.

Change-Id: Ib9e487993a92a0d71daef32c4f61d01d812869a7
Signed-off-by: Krzysztof Karas <krzysztof.karas@intel.com>