lib/idxd: fix device reference counting in kernel IDXD
Context created by accfg_new() needs to be kept during
the configuration and usage of the DSA device.

The creation of the context increments the reference count,
which can be further managed by accfg_ref() and accfg_unref().

Before this patch, the context was created correctly but
was not assigned to kernel_idxd->ctx, which is used
in kernel_idxd_device_destruct(). This issue is now resolved.

This patch ensures that the system can handle any number of devices,
or no device at all. To accommodate this, an additional reference
is taken for each attached device. Once the configuration is complete,
the context reference used for configuration is released.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Change-Id: I5e44867bfe3ed64f98ead61551568c0d90d185a0