lib/idxd: free batches on channel creation failure
Batches were freed on failure during allocation,
and when channel was put back. Any other failures
during creation of the channel were not accounted for.

This patch adds a common function to free batches on the channel,
and uses it on error path for channel creation.
Specifically when _vtophys() fails, previously the batches
might have not been freed during spdk_idxd_get_channel.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Change-Id: I61e1cff745e3c6eeeea3df8b2b67a3009b95c748