lib/idxd: track WQ use on single descriptor granularity
Before the use of WQ was tracked by splitting it in chan_per_device
parts, and utilizing each one for a channel.

This patch now changes it so that each entry in the WQ is separate,
and can match to any channel that requested it.
The number of descriptors per channel still remain the same,
but later in the series it will be configurable or maybe even
dynamic in some capacity.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Change-Id: I3f648c2e36fbad8e976b5d353e405eba5cfabf1a