lib/idxd: refactor to single batch for each descriptor
Number of batches for each channel were already tied to
the number of descriptors. This patch makes it so that
each batch is allocated at the same time as each descriptor.

All existing batches on a channel are either on batch_pool
list or in chan->batch - which points to current batch being
prepared for submission.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Change-Id: I5a1a7ee9a5c2d1e5d4d07743fee64340d96ce95b