event: do not call reactor events from spdk_thread context
There are no remaining events that require running in an spdk_thread
context, so remove that code so that spdk_events are explicitly *not*
called from an spdk_thread context.

This is the desired behavior, there should be a clear delineation between
reactor context and spdk_thread context, and this change ensures that we
can send events that will not be run in an spdk_thread context.

So now spdk_reactor_set_interrupt_mode() checks that it was called on the
scheduling reactor, and not from the context of an spdk_thread.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Signed-off-by: syeon.shin <syeon.shin@samsung.com>
Change-Id: Iadfae6f0943ad77076ffb8e81e5154d998e7a6f4