lib/iscsi: change iscsi conn schedule strategy
In a production environment, it is common to allocate
one CPU core to each iSCSI target. However, during testing,
it was observed that when there are four targets and four cores
in SPDK, two targets were running on the same core.

Further investigation revealed that the issue was occurring
in the iscsi_conn_schedule function, where multiple targets
were being assigned to the same poll group (pg) when the number
of CPU cores matched the target usage.

+---+        +-------+     +-------+
|Ini|        |Target1|     |Target2|
+---+        +-------+     +-------+

        pg1
s1: ------------>

                pg2
s2: --------------------------->
s3: <--------------------------

                pg1
s4: --------------------------->

The above scenario involves two targets and two poll groups.
When the first connection is established with target1, it is assigned to pg1.
Meanwhile, a new connection is established with target2. In this case,
target2 is assigned to pg2. However, when the first connection with target2
is closed, and another connection is still established with target2, the previous
logic assigns pg1 to target2. This results in one poll group handling multiple
targets when the number of poll groups matches the target usage.

To address this issue, the scheduling strategy has been modified to make
the target always use the pg it uses for the first time.

Change-Id: Ied720c28ebdc72247f3567fc77663ecbd71ffe59
Signed-off-by: Changqi Lu <luchangqi.123@bytedance.com>
Signed-off-by: zhenwei pi <pizhenwei@bytedance.com>