[RFC] test/nvmf: add SPDK target vs kernel initiator test
Previously it was difficult to test the block device that would be
added after 'nvme connect' due to stability issues (the device would
not appear or it would pop-in and then disappear).

After commit 50abcc179e0c ("nvme-tcp: strict pdu pacing to avoid send
stalls on TLS") has been merged on the kernel side, the stability has
improved and we now can run IO on the block device.

NOTE: currently this is present on kernel 6.10.0-rc1, so until we
upgrade to 6.10 in the future, this patch cannot be merged.

Change-Id: I15f8e7aff9b04c6b503491daa44f901a260382e2
Signed-off-by: Krzysztof Karas <krzysztof.karas@intel.com>