bdev: [RFC] no_metadata flag for bdev
Added a bdev open flag, to control the block size
exposed by the bdev. no_metadata set to true will configure
the bdev to expose the block size equal to data size, without
the metadata, even if the underlying bdev is formatted with
metadata.

If needed, the generic bdev layer will add the metadata
and generate/remove the DIF/DIX if required.

For now, the async version of bdev open has been extended.
Future patches will add this flag to a sync version of bdev
open.

Change-Id: I66e39a0c225042c964cf0de5e92c5195aeaba25a
Signed-off-by: Slawomir Ptak <slawomir.ptak@intel.com>