lib/ublk: add option to retain ublk device on exiting or recovery fails
Currently, the ublk device is actively closed by SPDK upon exit
or recovery failure. This can lead to issues when the ublk device
is being used by other processes, potentially preventing the ublk
device from being properly closed. Additionally, in scenarios such
as upgrades, users prefer to recover the device rather than
closing and restarting it. So, support is added to prevent the
ublk device from being actively closed on exit or recovery failure,
allowing the user to decide when to actually close the ublk device.

Change-Id: Ibd0c3dfb86e4c1451b699e8ccd21eb82ebdd0bdd
Signed-off-by: Haolin Liu <hollinisme@gmail.com>