bdev: [RFC] Introduce a no_metadata flag
The no_metadata flag can be set on a per IO basis, and if set,
the application can send data buffers without metadata,
even if the underlying bdev uses extended block size. The generic
bdev layer will apply the necessary conversion, which may include
allocating metadata buffers and/or generating/removing DIF/DIX.

This flag needs to be carried through the bdev IO stack, to the point,
where Accel DIF/DIX operations can be appended to the Accel sequence.
This append needs to happen at the right time, since DIF/DIX needs to
be calculated on the data, which will not be modified by any other
Accel operations in that sequence.

Change-Id: If63b94e6f65938bbf2bced1aaec921dbd850c854
Signed-off-by: Slawomir Ptak <slawomir.ptak@intel.com>