lib/idxd: refactor chan_per_device to descriptors_per_channel
In few of the following patches, it will be possible to configure
number of descriptors assigned for each channel via RPC.
Idea is to be more flexible with the number of possible channels
to create out of the total_wq_size (and multiple IDXD devices).

In this patch just replace the chan_per_device with descriptors_per_channel
field and set it during workqueue setup.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Change-Id: I748f6dfaec5f3c7d47400609aad4e7c5ed1ed8ec