lib/blob: add spdk_blob_get_bs()
Add a function to get blobstore from blob. Having a pointer
to blob, the blobstore should always be available.

This will be useful in following series, to get blobstore
pointer when iterating over blobs in spdk_bs_load().
Since that pointer usually is not available until spd_bs_load()
finishes.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Change-Id: Ie0fc76b17b04857979a6ed7e9f21675fa8fee171