lib/lvol: utilize the iter_cb_fn during lvs load
This patch adds the iter_cb_fn when loading the lvolstore.
For now all it does is adds debug prints.

Later, it will be possible to read xattrs from super blob or
open lvols immediately. Instead of waiting until lvs finishes loading.
This will help with issue #3359 and in general simplify the lvs load.

Note that spdk_bs_get_super() is executed immediately, so we can
expect to get the super_blob_id while still in context of iter_cb_fn.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Change-Id: I4697baa3608f5a8c95a122b722a0c98911e95479