thread: remove pool parameter from spdk_iobuf_for_each_entry
We always want to iterate across the entire iobuf_channel, so just
make the iterator do exactly that, rather than requiring the caller
to have to iterate both of the pools itself.

This both simplifies the calling logic, and prepares for upcoming
changes which will support multiple NUMA node caches in one channel.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: Ieed144671e7ee7cb4d7b7b28e803ea3cae641fee