module/dsa: allow both driver types
If config_kernel_mode is not specified via RPC (optional),
then attach both driver types for DSA devices.

To achieve that new enum and decoder function was added to
handle the case when this key is not present in JSON.

Any user that explicitly passed this option is JSON config
in the past, will still get only specific driver type.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Change-Id: I675f601bfe65b57db2f688a75390b92f0ea84459