[RFC]module/dsa: add descriptors_per_channel to RPC
This is a proof of concept for changes that should allow
to define this value on per module (or device) level.

The spdk_idxd_device is opaque pointer, so what is here will
not work. Instead controller opts similar to NVMe driver should
be added - set by caller in probe_cb() and used when device
is attached.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Change-Id: I14ab0d97ff8f9f3f2804bc229d51284eafef588d