thread: make spdk_iobuf_entry_abort() find its entry in the STAILQ
This prepares for upcoming changes where we will have multiple queues
to search, one per NUMA-node. We won't know which NUMA-node's queue
the entry will be found on.

We could theoretically store the numa_id in the spdk_iobuf_entry,
but that increases its size and requires other changes in existing
structures that include it. Since this is a very uncommon operation,
walking the STAILQs is fine.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: I645fc6d2dab815ff70163932e0e8a45953780113