fio_plugin/bdev: remove spdk_fio_cleanup_thread
Now the only thread we need to cleanup in spdk_init_thread_poll is the
init thread it self, and we just need to add it to the g_threads list to
exit it along with threads created by spdk modules. So just replace
spdk_fio_cleanup_thread with a single tailq insert operation.

Lock should not be needed at this time.

Change-Id: I4bbf011846357496b7a5fac82611f9bc13869eec
Signed-off-by: Jinlong Chen <chenjinlong.cjl@alibaba-inc.com>