[TEST] nvmf/rdma: use private_data cntlid to steer qpair to bdev numa_id
This is only a prototype.

RDMA connections for I/O queues pass the intended cntlid in the
RDMA_CM private data. We can encode the subsystem ID into the
generated cntlids, such that on a connect, we can pick the optimal
core based on the NUMA ID of the bdev in the targeted subsystem.

Note that the subsystem is not explicitly stated until the initiator
sends the FABRICS_CONNECT command, which happens after a core has
been chosen for the RDMA connection. So we need to use this
hint/hack using the ctrlid to know the subsystem sooner.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: I7072cd01ca5ebceff2af4d21053d388aa1e3aa8d