test/nvmf: Solve ambiguity around $NVMF_SECOND_TARGET_IP
Its usage is confusing, depending on the transport that is being
tested. Logically, initiator + target setup is meant to emulate
separate nodes on the network. For tcp this is quite clear since
regardless of the underlying platform (virt or phy) we construct
a separate namespace for the target device.

For rdma, however, it's not that clear. When tests are running
under rdma, some assumptions are made - first, that all of them
are executed under actual phy setup (virt was dropped some time
ago) and that all supported devices come from a single NIC.
This is similar to tcp + phy, but the difference is within the
namespace - infiniband devices that we pick up, are kept in the
main namespace - and the fact that these devices don't have to
be connected over loopback to be in UP state.

If test under tcp is using $NVMF_*_TARGET_IP, we know that it
references a "separate" node on the network. But for rdma,
this is obviously not the case.

Since setup_interfaces() constructs initiator + target pairs,
for rdma we can treat the "initiator" interface as an actual
target instead.

Change-Id: I0fed2be1df704029314a0f94687e1fa6678f4046
Signed-off-by: Michal Berger <michal.berger@intel.com>