nvme: enable interrupts for pcie nvme devices
Add an option enable_interrupts to the spdk_nvme_ctrlr_opts structure.
If this is set to true for pcie controllers interrupts may be enabled
during initialization. Applications are required to check the resulting
value after the attach step to check for success.
Maximum of 256 eventfds can be reserved for I/O queues, but the actual
number can be lower and is based on the minimum requested I/O queues
and number of available I/O queues.
The nvme_pcie_ctrlr_cmd_create_io_cq() interface has been modified to
create I/O completion queues with interrupts. The interrupt vector field
corresponds to the queue identifier in this case.

This is only supported within a primary SPDK process, and if enabled
SPDK will not support any secondary processes.

Change-Id: Iff4e2348a0b77199cabb30c0bf93e0eed920cc93
Signed-off-by: Ankit Kumar <ankit.kumar@samsung.com>