scripts/core-collector: Reroute core dumps to container's namespace
Our CI executes several jobs with different set of tests (e.g. ut)
under containers. Normally this is fine, but there's a small issue:
core-collector, when kernel pipes core to it from a crashing process,
is always executed within the main namespace. This means that core
lands "outside" our workspace.

Add support for looking up crashing process's namespace and rerouting
core into the container.

This still requires the "host" side to use our core-collector as we
are not able to modify procfs entries from within the container.
Nonetheless, it should bring the awareness of a potential issue that
we may be facing.

Change-Id: I0ebf2012e9ddf9467aeeaef08b44d98c47902f65
Signed-off-by: Michal Berger <michal.berger@intel.com>