[TEST] Add concept of "pending_work" to pollers and threads
This is part of investigations into using this metric to determine
if and when we should either move spdk_threads as part of scheduling
or drop a reactor into interrupt mode.

The concept is that it is better to only move spdk_threads or
go into interrupt mode when things are truly idle. Interrupt mode
for example has a full 3-4us transition period, which impacts
performance even at low queue depths.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: If3fddf9a7aec2d76ee4a77d26f3403b12279c6a5