nvme/poll_group: create and manage fd_group for nvme poll group
This commit adds spdk_fd_group to spdk_nvme_poll_group, which manages
interrupt events for all the file descriptors of spdk_nvme_qpair that
are part of this poll group.

Four new APIs have been added which creates fd group and manages events
on the fds registered in the fd group.

1). spdk_nvme_poll_group_create_fd_group()
Creates a new spdk_fd_group for the nvme poll group. This is to manage
events on the fds of all the nvme qpairs in the poll group, collectively.

2). spdk_nvme_poll_group_get_fd()
fetches the internal epoll_fd of fd_group within the nvme poll group.

3). spdk_nvme_poll_group_add_qpair_fd_ext()
Registering event source on a file descriptor for a spdk_nvme_qpair to
the fd_group of spdk_nvme_poll_group is the tricky part since event
handler requires a callback function and an argument to handle events.
For this case these need to be a function which processes the spdk nvme
qpair and an argument which encapsulates the spdk nvme qpair.
Example: In case of nvme bdev it needs to be the nvme_qpair structure.

At the time of spdk_nvme_poll_group creation we are unaware of these.
Hence we cannot add a spdk_nvme_poll_group_create_ext() version where we
save the above mentioned cb_fn, arg and a flag which could have been
used during spdk_nvme_poll_group_add() or
spdk_nvme_ctrlr_connect_io_qpair() to register event source for the
spdk_nvme_qpair fd to spdk_nvme_poll_group.
Then there are various transports which may register different event
types and can have different fd types.

Taking all these into consideration the best way forward is to introduce
an API which expects a function and an argument along with extended
event handler opts which specifies the events and the fd type. The event
source for file descriptor of the spdk_nvme_qpair is registered to the
epoll_fd of the fd_group within the spdk_nvme_poll_group.

4). spdk_nvme_poll_group_wait(), collectively waits for events on all the
spdk_nvme_qpair file descriptors managed in the fd_group of poll group.
In case of events the function with argument gets called. This is the
same function and argument which was specified in
spdk_nvme_poll_group_add_qpair_fd_ext().

No new API has been exposed to unregister event source on file descriptor
of spdk_nvme_qpair from the fd_group within the poll group. This will
automatically happen whenever qpair disconnect gets called.

Similarly no new API has been exposed to destroy the fd group within the
poll group. This will be taken care by spdk_nvme_poll_group_destroy().

Change-Id: If40f1ea82051ae598590f5a23ab9ed58bcb4af09
Signed-off-by: Ankit Kumar <ankit.kumar@samsung.com>