test/nvme/perf: Check for NVMe devices before creating logical volumes
Modified the create_lvols() function to check for the presence
of NVMe devices before attempting to create logical volumes
when the --use-lvol-bdevs option is specified. If no NVMe devices
are found, the script exits with an error. This ensures that the
test run configuration matches the requested CLI options by
confirming the availability of NVMe devices before running the
performance tests.

Change-Id: Ib7c2fea3f3a9d054869f4b2584096b2d30bace28
Signed-off-by: Jaroslaw Chachulski <jaroslawx.chachulski@intel.com>