util: fix total fds to wait for
The epoll file descriptor for a fd group waits for interrupt event on
the fds in its own fd group as well as the fds in all its children fd
group. Add a new field num_descendant_fds to the structure
spdk_fd_group that keeps track of all its children fd group fds.
Use this to calculate the actual number of fds to wait for a fd group
in spdk_fd_group_wait(), which is num_fds + num_descendant_fds.

Change num_fds type from signed to unsigned int.

Change-Id: Ib06de08e1083579540f9d0d0ee5fb8db23caad11
Signed-off-by: Ankit Kumar <ankit.kumar@samsung.com>