test/app: Create NVMF target testing application using bdev_nvme module
This change creates an SPDK application which acts as a sample SPDK NVMF
initiator that uses bdev_nvme module to test discovery, connect and
perform IO against a NVMF target. Subsequently will also add tests for
PR and multipath path of the target.

Motivation
SPDK currently has an application for identifying NVMF targets
(spdk_nvme_identify), but it does not test IO or PR operations.
Additionally, we wanted to use the multipathing capabilities provided
by the bdev_nvme module, and found that they are not being used outside
the scope of RPCs currently. We aim to develop such an application that
includes these features to verify that the target is functioning
correctly.

Moreover, we found an issue earlier in spdk while using Dell Powerflex
target, which could have been caught if we had such a testing application.

Implementation
The application uses the bdev_nvme module to leverage discovery and
connectivity features and create SPDK bdevs that support multipathing.
After bdev creation is complete, an spdk_write operation is followed by
an spdk_read operation to test IO. The application currently only tests
for discovery, connect and IO. Later patches will add support for PR
and multipath testing to ensure full compatibility.

Change-Id: I66e32443b7ad0e942a2c5083711b54982fbc86a9
Signed-off-by: Shravan Garg <shravan.garg@nutanix.com>