lib/nvme: callback function to manage events
Registering event source on the file descriptor of spdk_nvme_qpair is
tricky when the qpairs are managed within the poll group. This poses a
challenge of notifying the application when an event gets generated.

Two new fields event_cb_fn and event_cb_arg are added to
spdk_nvme_io_qpair_opts. The event_cb_fn is a callback function which
can process the qpair, and event_cb_arg is an argument which encapsulates
the qpair. If event_cb_fn is specified without specifying event_cb_arg,
then event_cb_arg is set to the newly created qpair address. This is done
for those application that require qpair as the callback argument when
events are generated.

Change-Id: Ib5cae25f67552bb4bd4bf9ce4c69a5d7999d8eba
Signed-off-by: Ankit Kumar <ankit.kumar@samsung.com>