examples: remove unnecessary thread_is_idle check in nvmf and thread_ex
Related to GitHub issue #3545. Checking spdk_thread_is_idle before
spdk_thread_destroy may cause confusion and imply that a thread must be
idle before it can exit, which would cause a thread with timed pollers
to hang when exiting. This is not the case; a thread does not need to
wait for timed pollers to expire before exiting. Cleaning up these
examples will improve clarity.

Change-Id: I16aff7eb7c27aa6bc20d76b1894312b520717a7f
Signed-off-by: Anisa Su <anisa.su@samsung.com>