test/unit/blob: use io_unit_size in more places
When blobstore was first created, we had concept of "page size". This
size was used both as the block size of blobs (i.e. the units that
users could read/write) and the metadata page size. They were both
hardcoded to 4096.

"io unit size" was introduced a long time ago. This allowed IO in
sizes different than 4096, usually 512. spdk_bs_get_io_unit_size()
can be used to get this value.

We are also working on support for metadata pages sizes greater than
4096. spdk_bs_get_page_size() will continue to return the size of a
metadata page, it may just not always be 4096.

But the unit tests in a lot of places use "page size" when "io unit
size" should be used instead. Many of the unit tests use 4096 as the
default blocklen, so everything still worked. But let's change
these to use io_unit_size correctly now, so that they will still pass
when the large IU metadata page sizes are introduced.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: I81d9d9b711a20e84bb31181dd167d2f7916500e0