test/unit/blob: do not make assumptions about first data cluster
A bunch of blobstore unit tests make assumptions that the first
data cluster is cluster index 1 (meaning cluster index 0 contains
all of the metadata).

But to support variable metadata page sizes, this assumption is no
longer valid. For example, with 16KB metadata page sizes, the
metadata will take up more than 1 cluster.

So improve the unit tests to remove assumptions in this area.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: If5112b8bcb7124cadc292dd1fde466cbf3d91cd4