test/unit/blob: add SZ to represent cluster size in blocks for io tests
The various test_io/test_iov helper functions are littered with
hard-coded values based on a known size of 32 blocks per cluster. We
will need to increase this size in an upcoming patch, to support
larger metadata page sizes. But instead of replacing all of the
existing hard-coded values with new hard-coded values, introduce
local SZ variables, and do various math based on that. Then increasing
the size of these clusters in the later patch will be trivial.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: Ie931b026e7b274854fbde492f8a785d5a03b5398