test/nvme/xnvme: Tidy the test suite
Split common stuff into common.sh which will try to emulate some sort
of a logic for the main tests which should resolve around different
io_mechanisms - the idea is to generally remove all the superfluous
checks and have a prep dedicated for each mechanism.

The xnvme_to_malloc_dd_copy test is removed altogether since in
context of xnvme it doesn't add anything sensible to the table -
it's removal won't affect the final code coverage.

Change-Id: If91cc6f2d3f6937fef22c534c57b2972caf26e8b
Signed-off-by: Michal Berger <michal.berger@intel.com>