test/nvme/xnvme: Enable polling in nvme driver
Under latest linux kernels (>=6.11.x) the underlying filename must
support polled io for the uring's IOPOLL to work. Since we use
nvme for these tests, make sure number of polling queues is
> 0 while setting up the driver.

in blockdev.sh xnvme is not the critical part of the tests so instead
of playing with nvme driver, we simply enable conserve_cpu which skips
IOPOLL setup altogether.

Change-Id: I63771c63a8be8bc3415c098ffeefa4475b8dc7ed
Signed-off-by: Michal Berger <michal.berger@intel.com>