bdev/malloc: Fix unexpected DIF verification error for initial read
The patch
https://github.com/spdk/spdk/commit/e1f15ba5418b38056fdb6d67186075aa1fada9e7
set reference tag to 0xFFFFFFFF and set application tag to 0xFFFF.

However, these are actually written to DIF fields only if their checks
are enabled. Usually, reference tag check is enabled but application tag
check is disabled.

For DIF type 1 or 2, DIF check is disabled only if application tag is
0xFFFF.

Hence, the value of 0xFFFFFFFF in the reference tag caused verification
error for any read.

The previous patch fixed DIF library to set application tag to 0 if its
check is disabled.
This patch fixes another problem, to set application tag to 0xFFFF as
source code comment says even if application tag check is disabled.

Signed-off-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Change-Id: I1d444eb6cc6de2e36abc7b0b85364c11dbe1df67